Puuduva sisendi korral jäi töövoog käima

+++ docker_start.sh
@@ -4,6 +4,9 @@ if [ ! -f /config/config.js ]; then
     cp -R -u -p /src/config_dist.js /config/config.js
 fi
 
-ln -s /config/config.js /src/config.js
+if [ ! -f /src/config.js ]; then
+    ln -s /config/config.js /src/config.js
+fi
+
 forever start /src/app.js
 forever list
